requirements.txt
streamlit==1.36.0
networkx==3.3
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
tensorflow==2.16.1
python-dotenv==1.0.1
pyyaml==6.0.1
streamlit-webrtc==0.47.4
scikit-learn==1.4.2
Pillow==10.3.0
streamlit-drawable-canvas==0.9.3